# Conflicts:
#	DiscordBot/Services/DatabaseService.cs

+++ DiscordBot/Domain/Casino/Games/RockPaperScissors/RockPaperScissors.cs
@@ -74,13 +74,13 @@ public class RockPaperScissors : ACasinoGame<RockPaperScissorsPlayerData, RockPa
         return playerWins ? GamePlayerResult.Won : GamePlayerResult.Lost;
     }
 
-    public override long CalculatePayout(GamePlayer player, ulong totalPot)
+    public override long CalculatePayout(GamePlayer player, long totalPot)
     {
         return player.Result switch
         {
-            GamePlayerResult.Won => (long)totalPot - (long)player.Bet, // Winner gets the total pot minus their bet
-            GamePlayerResult.Lost => -(long)player.Bet, // Loser loses their bet
-            GamePlayerResult.Tie => 0, // Tie gets bet back (no loss, no gain)
+            GamePlayerResult.Won => totalPot - player.Bet,
+            GamePlayerResult.Lost => -player.Bet,
+            GamePlayerResult.Tie => 0,
             _ => 0
         };
     }
